+++ client/proxgui.h
@@ -22,7 +22,7 @@ void ShowGraphWindow(void);
 void HideGraphWindow(void);
 void RepaintGraphWindow(void);
 void MainGraphics(void);
-void InitGraphics(int argc, char **argv, char *script_cmds_file, char *script_cmd);
+void InitGraphics(int argc, char **argv, char *script_cmds_file, char *script_cmd, bool stayInCommandLoop);
 void ExitGraphics(void);
 #ifndef MAX_GRAPH_TRACE_LEN
 #define MAX_GRAPH_TRACE_LEN (40000 * 8)