+++ client/proxmark3.h
@@ -26,7 +26,7 @@ extern "C" {
 
 const char *get_my_executable_path(void);
 const char *get_my_executable_directory(void);
-void main_loop(char *script_cmds_file, char *script_cmd);
+void main_loop(char *script_cmds_file, char *script_cmd, bool stayInCommandLoop);
 
 #ifdef __cplusplus
 }