+++ CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to this project will be documented in this file.
 This project uses the changelog in accordance with [keepchangelog](http://keepachangelog.com/). Please use this to write notable changes, which is not the same as git commit log...
 
 ## [unreleased][unreleased]
+ - Add client option `-i` to stay in interactive mode after a script or command (@DidierStevens/@doegox)
  - Add VSCode tasks (@ViRb3)
  - Better warn user of hardcoded hitag info (@ViRb3)
  - Format and docs hitag (@ViRb3)